requirements.txt
asn1crypto==0.24.0
Authlib==0.8
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
codecov==2.0.15
coverage==4.5.1
cryptography==2.3
docopt==0.6.2
Flask==1.0.2
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
grip==4.5.2
gspread==3.0.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
path-and-address==2.0.1
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.2.0
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
six==1.11.0
spark-parser==1.8.7
SQLAlchemy==1.2.10
uncompyle6==3.2.3
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
xdis==3.8.6
